requirements.txt
# Core FastAPI and web dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6

# OCR dependencies (Tesseract-based)
pytesseract==0.3.10

# Image and PDF processing
Pillow==10.1.0
pdf2image==1.16.3
opencv-python-headless==4.8.1.78

# HTTP requests
requests==2.31.0
httpx==0.25.2

# Database connectivity
mysql-connector-python==8.2.0
pymysql==1.1.0

# JSON processing
demjson3==3.0.6

# Additional utilities
python-dotenv==1.0.0
pydantic==2.5.0
email-validator==2.1.0

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
